Fix naming convetions, removing duplicates, IObstacle
