+++ Assets/2DSprite/FlappyBird.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 564f4469858ed9742b505aee92a36e0b
+guid: 0a47d1a956422a44aa9ea7751c6729d6
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
