--- Assets/Scripts/IObstacles.cs
@@ -1,6 +1,0 @@
-
-public interface IObstacles
-{
-    public void MoveObstacle();
-    public void InitiateRandomPosition();
-}
