--- Assets/Scripts/IObstacles.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6940693f437fec048995b89738a52aa6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
